requirements.txt
pdfplumber
requests
Office365-REST-Python-Client